Write pre-computed 32-bit palette values directly into canvas pixel buffer
Instead of first writing 8-bit palette indices into an intermediate
buffer and then converting them in a second step into 32-bit values
for display.

This makes the raster process a bit slower than if we just wrote the
8-bit values during this first step, but the overall time to render is
faster. For applications with notable overdraw (same pixel written to
multiple times) this change could make the overall render time slower,
but otherwise it should be an improvement.